restructure scene code, not yet functional
